ulong taxid. fix worker dash recursive. fix list alignement

+++ src/exception/person_exception.cpp
@@ -5,7 +5,7 @@ PersonDoesNotExist::PersonDoesNotExist(const std::string& name, int tributaryNum
     std::logic_error(name + ", with number " + std::to_string(tributaryNumber) + ", does not exist!"){
 }
 
-PersonDoesNotExist::PersonDoesNotExist(int taxID) :
+PersonDoesNotExist::PersonDoesNotExist(unsigned long taxID) :
     std::logic_error("Person with number " + std::to_string(taxID) + " does not exist!"){
 }
